🎉 feat(jade.png): add new image file A new image file named jade.png has been added to the assets/images directory.
🎨 style(IndexPage.vue): update partner links with referrer parameter
The partner links have been updated to include a referrer parameter that specifies the source of the traffic. This will allow us to track the effectiveness of our partnerships and make data-driven decisions in the future.

🔀 chore(IndexPage.vue): update web3phl logo link
The web3phl logo link has been updated to include a referrer link to https://ossph.org. This will help track the source of the traffic to the web3phl website.

🎨 style(IndexPage.vue): adjust layout of images and add Jade's photo
The layout of the images was adjusted to make them more consistent and visually appealing. Jade's photo was also added to the list of images.